refactor: rename get_filepaths() to list_files(), use it in iac and sca
iac.filter and sca.file_selection do not need to create File instance, they
just work with paths. Make them use files.list_files() directly. All that
needed to be added was filtering out binary files.

Also move test_get_ignored_files() from verticals.sca.test_file_selection
to utils.test_files since the test actually does not depend of SCA.
