+++ ggshield/verticals/iac/filter.py
@@ -1,8 +1,7 @@
 from pathlib import Path
 from typing import List, Pattern, Set
 
-from ggshield.core.scan.file import get_files_from_paths
-from ggshield.utils.files import ListFilesMode
+from ggshield.utils.files import ListFilesMode, is_path_binary, list_files
 
 
 IAC_EXTENSIONS = {
@@ -32,7 +31,7 @@ def get_iac_files_from_path(
     :param verbose: Option that displays filepaths as they are scanned
     :param ignore_git: Ignore that the folder is a git repository. If False, only files added to git are scanned
     """
-    files, _ = get_files_from_paths(
+    paths = list_files(
         paths=[path],
         exclusion_regexes=exclusion_regexes,
         list_files_mode=(
@@ -45,10 +44,12 @@ def get_iac_files_from_path(
             )
         ),
     )
-    return [x.path for x in files if is_iac_file_path(x.path)]
+    return [x for x in paths if is_iac_file_path(x)]
 
 
 def is_iac_file_path(path: Path) -> bool:
+    if is_path_binary(path):
+        return False
     if any(ext in IAC_EXTENSIONS for ext in path.suffixes):
         return True
     if any(path.name.endswith(iac_ext) for iac_ext in IAC_EXTENSIONS):
